Reactive error handling implementation
