+++ src/main/java/com/nttdata/bootcamp/customerinfoservice/config/Constants.java
@@ -1,4 +1,4 @@
-package com.nttdata.bootcamp.customerinfoservice.utils;
+package com.nttdata.bootcamp.customerinfoservice.config;
 
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
