+++ src/main/resources/application.yml
@@ -1,13 +1,22 @@
-# App config
+# Server config
 server:
   port: 8081
 
-# Mongodb Config
+# Spring Config
 spring:
   data:
     mongodb:
       database: ProjectOne
       host: localhost
+  application:
+    name: customer-info-service
+
+eureka:
+  instance:
+    hostname: localhost
+    serviceUrl:
+      defaultZone: http://${eureka.instance.hostname}:8761/eureka/
+    prefer-ip-address: true
 
 # Constants
 constants:
